# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/jobs": {
      "post": {
        "operationId": "JobsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a draft job",
        "tags": [
          "Jobs"
        ]
      },
      "get": {
        "operationId": "JobsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "closed",
                "archived"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Job"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List and search jobs",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}": {
      "get": {
        "operationId": "JobsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a job by ID",
        "tags": [
          "Jobs"
        ]
      },
      "patch": {
        "operationId": "JobsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a job",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/publish": {
      "post": {
        "operationId": "JobsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a job",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/close": {
      "post": {
        "operationId": "JobsController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close a job",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/archive": {
      "post": {
        "operationId": "JobsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user lacks the required role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a job",
        "tags": [
          "Jobs"
        ]
      }
    },
...
  },
  "info": {
    "title": "GEDPro API",
    "description": "Applicant tracking API for users, candidates, jobs, applications, forms, documents, and interviews.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/v1",
      "description": "Version 1 (recommended)"
    },
    {
      "url": "/",
      "description": "Legacy unversioned compatibility"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "title should not be empty"
            ]
          },
          "path": {
            "type": "string",
            "example": "/jobs"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-15T10:00:00.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "req_01JABCDEF"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "failedLoginAttempts": {
            "type": "number"
          },
          "lockedUntil": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "role": {
            "enum": [
              "admin",
              "rh",
              "manager",
              "candidate"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "password",
          "isActive",
          "emailVerified",
          "failedLoginAttempts",
          "lockedUntil",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CandidateHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "candidate": {
            "$ref": "#/components/schemas/Candidate"
          },
          "previousState": {
            "type": "string",
            "enum": [
              "new",
              "preselected",
              "interview_scheduled",
              "in_interview",
              "accepted",
              "rejected"
            ]
          },
          "newState": {
            "type": "string",
            "enum": [
              "new",
              "preselected",
              "interview_scheduled",
              "in_interview",
              "accepted",
              "rejected"
            ]
          },
          "comment": {
            "type": "string"
          },
          "changedBy": {
            "$ref": "#/components/schemas/User"
          },
          "changedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidate",
          "previousState",
          "newState",
          "comment",
          "changedBy",
          "changedAt"
        ]
      },
      "Candidate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "normalizedEmail": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "normalizedPhone": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "archivedAt": {
            "format": "date-time",
            "type": "string"
          },
          "privacyConsent": {
            "type": "boolean"
          },
          "consentAt": {
            "format": "date-time",
            "type": "string"
          },
          "retentionUntil": {
            "format": "date-time",
            "type": "string"
          },
          "deletionRequestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "mergedInto": {
            "$ref": "#/components/schemas/Candidate"
          },
          "currentState": {
            "enum": [
              "new",
              "preselected",
              "interview_scheduled",
              "in_interview",
              "accepted",
              "rejected"
            ],
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateHistory"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "normalizedEmail",
          "phone",
          "tags",
          "skills",
          "privacyConsent",
          "currentState",
          "history",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
      "PipelineTransition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "fromStage": {
            "$ref": "#/components/schemas/PipelineStage"
          },
          "toStage": {
            "$ref": "#/components/schemas/PipelineStage"
          }
        },
        "required": [
          "id",
          "fromStage",
          "toStage"
        ]
      },
      "PipelineStage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "pipeline": {
            "$ref": "#/components/schemas/Pipeline"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "enum": [
              "applied",
              "screening",
              "interview",
              "offer",
              "hired",
              "rejected",
              "withdrawn"
            ],
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "archived": {
            "type": "boolean"
          },
          "outgoingTransitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineTransition"
            }
          }
        },
        "required": [
          "id",
          "pipeline",
          "name",
          "category",
          "position",
          "archived",
          "outgoingTransitions"
        ]
      },
      "Pipeline": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isTemplate": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineStage"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "isTemplate",
          "archived",
          "stages",
          "createdAt",
          "updatedAt"
        ]
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "employmentType": {
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "published",
              "closed",
              "archived"
            ],
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "pipeline": {
            "$ref": "#/components/schemas/Pipeline"
          },
          "applications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Application"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string"
          },
          "closedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "status",
          "owner",
          "pipeline",
          "applications",
          "createdAt",
          "updatedAt"
        ]
      },
      "ApplicationHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "application": {
            "$ref": "#/components/schemas/Application"
          },
          "previousStage": {
            "$ref": "#/components/schemas/PipelineStage"
          },
          "newStage": {
            "$ref": "#/components/schemas/PipelineStage"
          },
          "comment": {
            "type": "string"
          },
          "changedBy": {
            "$ref": "#/components/schemas/User"
          },
          "changedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "application",
          "newStage",
          "changedBy",
          "changedAt"
        ]
      },
      "Application": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "candidate": {
            "$ref": "#/components/schemas/Candidate"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "currentStage": {
            "$ref": "#/components/schemas/PipelineStage"
          },
          "source": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicationHistory"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "candidate",
          "job",
          "currentStage",
          "history",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
      "CreateJobDto": {
        "type": "object",
        "properties": {
          "pipelineId": {
            "type": "number"
          },
          "title": {
            "type": "string",
            "maxLength": 160
          },
          "description": {
            "type": "string"
          },
          "department": {
            "type": "string",
            "maxLength": 120
          },
          "location": {
            "type": "string",
            "maxLength": 160
          },
          "employmentType": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "pipelineId",
          "title",
          "description"
        ]
      },
      "UpdateJobDto": {
        "type": "object",
        "properties": {
          "pipelineId": {
            "type": "number"
          },
          "title": {
            "type": "string",
            "maxLength": 160
          },
          "description": {
            "type": "string"
          },
          "department": {
            "type": "string",
            "maxLength": 120
          },
          "location": {
            "type": "string",
            "maxLength": 160
          },
          "employmentType": {
            "type": "string",
            "maxLength": 80
          }
        }
      },
...
    }
  }
}